refactor(ops): move protocol/transform/ops → protocol/ops
The ops package provides protocol-level pure functions callable by any
layer (transform, server, client). Nesting it under transform/ implied
it was transform-internal, which was misleading and prevented callers
outside transform/ from importing it without a cross-package smell.

Moving to internal/protocol/ops/ makes the package a peer of transform/
at the protocol layer — the import path now reflects its actual scope.

All callers updated:
- internal/protocol/transform/{vendor,openai_cursor_compat}.go
- internal/client/{claude_client,claude_round_tripper}.go
- internal/server/protocol_roundtrip.go

No logic changes; package name stays "ops".
