+++ internal/protocol/ops/request_anthropic_model.go
