+++ internal/protocol/ops/request_anthropic_model_test.go
