+++ internal/protocol/ops/request_openai_codex.go
