+++ internal/protocol/ops/request_openai_extensions.go
