+++ internal/protocol/ops/request_openai_gemini.go
