+++ internal/protocol/ops/request_openai_gemini_test.go
