+++ internal/protocol/ops/response_openai_extensions.go
